# [WIP] Protobuf Performance Refactor