--- a/proto/marshal/collections/maps.py
+++ b/proto/marshal/collections/maps.py
@@ -14,6 +14,7 @@
 
 import collections
 
+import proto
 from proto.utils import cached_property
 
 
@@ -25,10 +26,18 @@ class MapComposite(collections.abc.MutableMapping):
     """
 
     @cached_property
+    def entry_class(self):
+        return self.pb.GetEntryClass()
+
+    @cached_property
     def _pb_type(self):
         """Return the protocol buffer type for this sequence."""
         # Huzzah, another hack. Still less bad than RepeatedComposite.
-        return type(self.pb.GetEntryClass()().value)
+        return type(self.entry_class().value)
+
+    @cached_property
+    def _override_pb_type(self):
+        return getattr(self.entry_class._meta["fields"]["value"], "pb_override", None)
 
     def __init__(self, sequence, *, marshal):
         """Initialize a wrapper around a protobuf map.
@@ -54,7 +63,10 @@ class MapComposite(collections.abc.MutableMapping):
         # buffers will create the key if it does not exist.
         if key not in self:
             raise KeyError(key)
-        return self._marshal.to_python(self._pb_type, self.pb[key])
+        obj = self._marshal.to_python(self._pb_type, self.pb[key])
+        if isinstance(obj, proto.Message):
+            obj._always_commit = True
+        return obj
 
     def __setitem__(self, key, value):
         pb_value = self._marshal.to_proto(self._pb_type, value, strict=True)