… proto/marshal/marshal.py
… proto/message.py
… tests/test_fields_composite.py
… tests/test_fields_map_composite.py
… tests/test_fields_optional.py
… tests/test_marshal_types_wrappers_bool.py