Aims to speed up naive usage on proto-plus-python as a wrapper around primitive protobuf objects.

This PR replaces attribute misses on proto-plus objects, which wound up in the `__getattribute__` override to query the hidden protobuf object directly, with descriptors which do the same inside a caching layer to optimize repeated access. This descriptor/caching layer also aims to avoid unnecessary marshaling altogether when circumstances allow, further reducing the runtime of basic tasks.

---

The following is a performance chart of several simple tasks executed against the same proto definitions, first as raw protobuf objects and again as a proto-plus object. The two definitions are:

_Note: All tasks were run 10,000 times._

| Generate a(n)... | pb2 | proto-plus (before refactor) | proto-plus (with refactor) | refactor impact |
|---|---|---|---|---|
| empty instance | 3ms | 10ms | 7ms | ⬇️ 30% 🟢 |
| populated instance | 9ms | 168ms | 118ms | ⬇️ 30% 🟢 |
| populated nested objs | 38ms | 211ms | 163ms | ⬇️ 23% 🟢 |
| populated nested objs (1x attribute reads) | 39ms | 326ms | 419ms | ⬆️ 29% 🔴 |
| populated nested objs (10x attribute reads) | 101ms | 2009ms | 706ms | ⬇️ 65% 🟢 |
| populated w serialization | 20ms | 195ms | 280ms | ⬆️ 43% 🔴 |
| populated w lots of fields | 25ms | 624ms | 440ms | ⬇️ 30% 🟢 |
| MapFields populated during init | 124ms | 799ms | 626ms | ⬇️ 21% 🟢 |
| MapFields populated after init | -- | 1079ms | 369ms | ⬇️ 66% 🟢 |
| deserializing objects | 116ms | 204ms | 180ms | ⬇️ 12% 🟢 |

---

When used as a dependency in python-datastore, the following proto-plus versions have these performance hydrating loaded 500 entities (network I/O is excluded):

| datastore version | notes | before googleapis/proto-plus-python#230 | with googleapis/proto-plus-python#230 |
|--|--|--|--|
| 1.15.3 | before proto-plus | 360ms | _N/A_ |
| 2.0.0 | with proto-plus | 675ms | 800ms |
| 2.1.2 | proto-plus usage optimized <br> via googleapis/python-datastore#155) | 550ms | 750ms |

---

Resolves #222