@software-dov 2021-05-17
Can you provide a link to your benchmark code? I've had a hard time creating good benchmarks that mimic real world message manipulation patterns.