@craiglabenz 2021-05-25
Closing this PR because it optimizes use cases that are different from what we see in client libraries like python-datastore and python-firestore.

Specifically, caching attributes is only helpful if they are read > 1 time, yet python-datastore converts all proto-plus instances into dictionary-like objects; meaning each field is read exactly once before all future reads are performed against that dictionary.

If you find this and would like to know more, reach out.